Don't crash when a chunk has a negative size
The chunk size is defined as a signed long and thus can be negative. This makes
little sense, but still we should return that the IFF is invalid instead of
segfaulting.

Instead of checking the chunksize directly, we check it at the malloc. This has
advantages for when there is another reason that the malloc fails, for example
if chunksize is very large.